refactor: move AlgoliaIndiceKey to types
