+++ src/lib/Scenes/Search/SearchModel.tsx
@@ -17,17 +17,6 @@ export interface SearchModel {
   clearRecentSearches: Action<SearchModel>
 }
 
-export enum AlgoliaIndiceKey {
-  Artist = "artist",
-  Article = "article",
-  Auction = "sale",
-  ArtistSeries = "artist_series",
-  Collection = "kaws_collection",
-  Fair = "fair",
-  Show = "partner_show",
-  Gallery = "partner_gallery",
-}
-
 export const getSearchModel = (): SearchModel => ({
   recentSearches: [],
   addRecentSearch: action((state, payload) => {
